+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Renamed `UpperCase` word formatter to `UpperCaseConverter`.
 - Renamed `ChangeDate` rule to `ChangeOnDateRule`.
 - Renamed `ChangeInterval` rule to `ChangeWithIntervalRule`.
+- Renamed `CharacterClass` rule to `CharacterClassRule`.
 
 ### Fixed
 - Nothing